Edit index.html and styles.css to have new css
I created a div for each of the recipes because I wanted them each to
be in their own row. I then edited the css so that each recipe has a
border box around them because my vision is to animate them so that
if the cursor is hovering over the recipe it'll extend out horizontally.